mv currencyConversions from appstore to lib
