+++ packages/app-store/hitpay/components/EventTypeAppSettingsInterface.tsx
@@ -1,15 +1,15 @@
 import { useState, useEffect } from "react";
 
 import type { EventTypeAppSettingsComponent } from "@calcom/app-store/types";
-import { useLocale } from "@calcom/lib/hooks/useLocale";
-import { Select } from "@calcom/ui/components/form";
-import { TextField } from "@calcom/ui/components/form";
-import { Alert } from "@calcom/ui/components/alert";
-
 import {
   convertToSmallestCurrencyUnit,
   convertFromSmallestToPresentableCurrencyUnit,
-} from "../lib/currencyConversions";
+} from "@calcom/lib/currencyConversions";
+import { useLocale } from "@calcom/lib/hooks/useLocale";
+import { Alert } from "@calcom/ui/components/alert";
+import { Select } from "@calcom/ui/components/form";
+import { TextField } from "@calcom/ui/components/form";
+
 import { paymentOptions, currencyOptions } from "./constants";
 
 const EventTypeAppSettingsInterface: EventTypeAppSettingsComponent = ({
