+++ packages/features/bookings/lib/handlePayment.ts
@@ -1,10 +1,10 @@
 import { PaymentServiceMap } from "@calcom/app-store/payment.services.generated";
 import type { EventTypeAppsList } from "@calcom/app-store/utils";
 import { eventTypeMetaDataSchemaWithTypedApps } from "@calcom/app-store/zod-utils";
-import type { AppCategories, Prisma, EventType } from "@calcom/prisma/client";
-import { convertToSmallestCurrencyUnit } from "@calcom/app-store/_utils/payments/currencyConversions";
 import type { Fields } from "@calcom/features/bookings/lib/getBookingFields";
 import { fieldTypesConfigMap } from "@calcom/features/form-builder/fieldTypes";
+import { convertToSmallestCurrencyUnit } from "@calcom/lib/currencyConversions";
+import type { AppCategories, Prisma, EventType } from "@calcom/prisma/client";
 import type { CalendarEvent } from "@calcom/types/Calendar";
 import type { IAbstractPaymentService } from "@calcom/types/PaymentService";
