+++ packages/lib/currencyConversions.ts
@@ -56,3 +56,16 @@ export const getCurrencySymbol = (currencyCode: string): string => {
     return "$";
   }
 };
+
+export const formatPrice = (price: number, currency: string | undefined, locale = "en") => {
+  switch (currency) {
+    case "BTC":
+      return `${price} sats`;
+    default:
+      currency = currency?.toUpperCase() || "USD";
+      return `${Intl.NumberFormat(locale, {
+        style: "currency",
+        currency: currency,
+      }).format(convertFromSmallestToPresentableCurrencyUnit(price, currency))}`;
+  }
+};
